+++ bin/flutter_pseudolocalizor.dart
@@ -5,9 +5,10 @@ import 'package:flutter_pseudolocalizor/flutter_pseudolocalizor.dart';
 import 'utils/yaml_parser.dart';
 
 void main() {
-  final packageSettings = YamlParser.packageSettingsFromPubspec();
+  final packageSettings = YamlParser.packageSettingsFromPubspecOrOwnYaml();
   if (packageSettings == null) {
-    print('Error! Settings for flutter_pseudolocalizor not found in pubspec.');
+    print(
+        'Error! Settings for flutter_pseudolocalizor not found in pubspec.yaml or flutter_pseudolocalizor.yaml.');
     exit(0);
   }